refactoring(test): refactoring unit tests
