+++ jrcc-access-spring-boot-autoconfigure/src/test/java/ca/gov/bc/open/jrccaccess/autoconfigure/plugins/rabbitmq/StandaloneAccessConfigurationTester.java
@@ -1,4 +1,4 @@
-package ca.gov.bc.open.jrccaccess.autoconfigure;
+package ca.gov.bc.open.jrccaccess.autoconfigure.plugins.rabbitmq;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -12,42 +12,38 @@ import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 
+import ca.gov.bc.open.jrccaccess.autoconfigure.AccessApplication;
+
 @RunWith(SpringRunner.class)
 @SpringBootTest(
         classes = AccessApplication.class,
         properties = {
-        		"spring.redis.cluster.nodes=127.0.0.1:5000,127.0.0.1:5001",
-        		"bcgov.access.input=http",
         		"bcgov.access.input.plugin=http",
         		"bcgov.access.output.plugin=rabbitmq"
         })
 @ContextConfiguration
-public class ClusterAccessConfigurationTester {
+public class StandaloneAccessConfigurationTester {
 
 	@Autowired
 	private JedisConnectionFactory jedisConnectionFactory;
-	
+
 	@Autowired
 	private CacheManager cacheManager;
-	
+
 	@Test
 	public void with_default_config_should_return_a_valid_stringRedisTemplate() {
 
 		assertNotNull(cacheManager);
-
 	}
-	
+
 	@Test
 	public void with_default_config_should_return_a_valid_jedisConnectionFactory() {
 
-		assertEquals(2, this.jedisConnectionFactory
-				.getClusterConfiguration()
-				.getClusterNodes().size());		
-		
-		this.jedisConnectionFactory
-				.getClusterConfiguration()
-				.getClusterNodes()
-				.forEach(redisNode -> assertEquals("127.0.0.1",redisNode.getHost()));		
+		int expectedPort = 6379;
+
+		assertEquals("localhost", this.jedisConnectionFactory.getHostName());
+
+		assertEquals(expectedPort, this.jedisConnectionFactory.getPort());
 	}
-	
+
 }
