+++ jrcc-access-spring-boot-autoconfigure/src/test/java/ca/gov/bc/open/jrccaccess/autoconfigure/plugins/rabbitmq/TopicExchangeTester.java
@@ -1,12 +1,13 @@
-package ca.gov.bc.open.jrccaccess.autoconfigure.rabbitmq;
+package ca.gov.bc.open.jrccaccess.autoconfigure.plugins.rabbitmq;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
-import org.springframework.amqp.support.converter.MessageConverter;
+import org.springframework.amqp.core.TopicExchange;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -23,16 +24,18 @@ import ca.gov.bc.open.jrccaccess.autoconfigure.AccessApplication;
 				"bcgov.access.output.plugin=rabbitmq"
 		})
 @ContextConfiguration
-public class MessageConverterTester {
+public class TopicExchangeTester {
 
-	@Qualifier("jsonMessageConverter")
+	@Qualifier("documentReadyTopic")
 	@Autowired
-	private MessageConverter sut;
+	private TopicExchange sut;
 	
 	@Test
-	public void with_default_should_return_jackson_converter() {
+	public void with_default_should_return_document_ready_topic() {
 	
-		assertEquals(Jackson2JsonMessageConverter.class, sut.getClass());
-
+		assertEquals("document.ready", sut.getName());
+		assertTrue(sut.isDurable());
+		assertFalse(sut.isAutoDelete());
+	
 	}
 }
